Endrer oppsett for hjemmekontor
